feat: update dependencies to latest versions
One thing to note is that Poisson in rand_distr now returns a floating
point value. Upon inspection, the team working on rand decided that this
makes for a cleaner interface, and thus ask that users cast the floats
to integers. This isn't the most beautiful approach, but it should be
sufficient.

+++ src/univariate/poisson.rs
@@ -57,7 +57,7 @@ impl Distribution for Poisson {
 
         rand_distr::Poisson::<f64>::new(get_lambda!(self))
             .unwrap()
-            .sample(rng)
+            .sample(rng) as u64
     }
 }
